Migrate skate-park and product-listing to CSDK v2.0.0
- Rename src/middleware.ts to src/proxy.ts and update to new Proxy API:
  replace *Middleware classes/defineMiddleware with *Proxy/defineProxy
  from @sitecore-content-sdk/nextjs/proxy; update function signature to
  export default function proxy(req: NextRequest) with no NextFetchEvent
- Fix src/app/api/sitemap-llm/route.ts: update SiteResolver and
  SitemapXmlOptions imports from core/* to content/* following the
  decoupling of @sitecore-content-sdk/content from core in v2.0.0
- Remove deprecated npm config block (appName, graphQLEndpointPath,
  language, template) from package.json in both projects
- Bump nodeVersion to 24 in xmcloud.build.json for both rendering hosts
  as CSDK v2.0.0 requires Node.js >=24
