+++ examples/kit-nextjs-product-listing/package.json
@@ -3,12 +3,6 @@
   "description": "Application utilizing Sitecore JavaScript Services and Next.js",
   "version": "1.0.0",
   "private": true,
-  "config": {
-    "appName": "kit-nextjs-product-starter",
-    "graphQLEndpointPath": "/sitecore/api/graph/edge",
-    "language": "en",
-    "template": "nextjs"
-  },
   "author": {
     "name": "Sitecore Corporation",
     "url": "https://doc.sitecore.com/xmc/en/developers/content-sdk/index.html"
