+++ examples/kit-nextjs-skate-park/package.json
@@ -2,12 +2,6 @@
   "name": "content-sdk-nextjs-skate-park-app-router",
   "version": "1.0.0",
   "private": true,
-  "config": {
-    "appName": "content-sdk-nextjs-skate-park-app-router",
-    "graphQLEndpointPath": "/sitecore/api/graph/edge",
-    "language": "en",
-    "template": "nextjs"
-  },
   "author": {
     "name": "Sitecore Corporation",
     "url": "https://doc.sitecore.com/xmc/en/developers/content-sdk/index.html"
